Infer download/upload directly based on source/sink; drop list/move/delete
Using an operation field to switch channel role diverged from
caterpillar's position-based model. The task now infers its role from the
channels like the file task - no input -> download (source), input -> upload
(sink). list/move/delete are dropped; auth, host-key
and retry helpers merged into sftp.go.
