+++ internal/pkg/pipeline/task/sftp/README.md
@@ -6,24 +6,21 @@ The `sftp` task transfers files to and from SFTP servers. It reads credentials f
 
 The `sftp` task handles SFTP only. It does not talk to S3 directly. Instead, it is combined with the existing [`file`](../file) task:
 
-- **Upload (S3 → SFTP)**: the `file` task reads from `s3://…` and the `sftp` task (`operation: upload`) writes the files to the server.
-- **Download (SFTP → S3)**: the `sftp` task (`operation: download`) reads files from the server and the `file` task writes them to `s3://…`.
+- **Upload (S3 → SFTP)**: the `file` task reads from `s3://…` and the `sftp` task writes the files to the server.
+- **Download (SFTP → S3)**: the `sftp` task reads files from the server and the `file` task writes them to `s3://…`.
 
 This reuses Caterpillar's existing S3 code. The file name passes from one task to the next through a record context value, so `file` and `sftp` work together without extra configuration.
 
 ## Behavior
 
-The `operation` field sets what the task does:
+Like the `file` task, the role is **inferred from the channels**:
 
-| Operation | Role | Input | Output | What it does |
-|-----------|------|-------|--------|--------------|
-| `upload` | sink | required | optional (pass-through) | Writes each record's data to the server. If `remote_path` is a directory, the source file name is added to it. |
-| `download` | source | none | required | Reads file(s) at `remote_path` (a single file, a glob, or a directory) and emits one record per file. |
-| `list` | source | none | required | Emits one JSON record per directory entry: `{name, size, mod_time, is_dir}`. |
-| `move` | action | optional | optional (pass-through) | Renames `remote_path` to `destination_path`. |
-| `delete` | action | optional | optional (pass-through) | Deletes `remote_path`. |
+| The task has… | Role | What it does |
+|---------------|------|--------------|
+| **no input** (it is the first task) | **source — download** | Reads file(s) at `remote_path` (a single file, a glob, or a directory) and emits one record per file. The base name is stored in the record context (`CATERPILLAR_FILE_NAME_WRITE`) so a downstream task can name what it writes. |
+| **an input** | **sink — upload** | Writes each incoming record's data to the server. If `remote_path` is a directory (trailing `/` or an existing directory), the source file name is appended. |
 
-For `move` and `delete`: if the task has an input, it runs once for each record (the paths can use record values). If it has no input, it runs once using the configured paths.
+It cannot be both: configuring the task with both an input and an output is an error.
 
 ## Authentication
 
@@ -61,7 +58,6 @@ If you set neither, the task refuses to connect (it fails closed). You can obtai
 |-------|------|---------|-------------|
 | `name` | string | - | Task name |
 | `type` | string | - | Must be `sftp` |
-| `operation` | string | - | One of `upload`, `download`, `list`, `move`, `delete` (required) |
 | `host` | string | - | SFTP server hostname (required) |
 | `port` | int | `22` | SFTP server port |
 | `username` | string | - | SFTP username (required) |
@@ -70,18 +66,18 @@ If you set neither, the task refuses to connect (it fails closed). You can obtai
 | `passphrase` | string | - | Passphrase for an encrypted `private_key` |
 | `host_key` | string | - | Authorized-key line used to verify the server |
 | `known_hosts_path` | string | - | Path to a `known_hosts` file |
-| `remote_path` | string | - | Remote file or directory (supports templating). On upload, a trailing `/` or an existing directory is treated as a directory. |
-| `destination_path` | string | - | Target path for `move` (supports templating) |
+| `remote_path` | string | - | Remote file or directory (required; supports templating). On upload, a trailing `/` or an existing directory is treated as a directory. |
 | `timeout` | duration | `30s` | SSH connection timeout (for example `15s`, `1m`) |
 | `max_retries` | int | `3` | Attempts per connect or transfer operation |
 | `retry_delay` | duration | `1s` | Delay between retries |
 | `fail_on_error` | bool | `false` | Stop the pipeline if this task fails |
-| `task_concurrency` | int | `1` | Number of parallel workers. For SFTP, each worker opens its own connection (see Notes and limitations) |
+| `task_concurrency` | int | `1` | Number of parallel workers. Each worker opens its own connection (see Notes and limitations) |
 | `context` | map | - | jq expressions that copy values from each record into its context for later tasks |
 
 ## Examples
 
 ### Upload files from S3 to an SFTP server
+The `sftp` task has an input (from the `file` task), so it acts as a sink and uploads.
 ```yaml
 tasks:
   - name: read_from_s3
@@ -89,7 +85,6 @@ tasks:
     path: s3://my-bucket/outbound/*.csv
   - name: push_to_client
     type: sftp
-    operation: upload
     host: sftp.client.example.com
     username: '{{ secret "/data/sftp/clientX/username" }}'
     private_key: |
@@ -99,11 +94,11 @@ tasks:
 ```
 
 ### Download files from an SFTP server to S3
+The `sftp` task has no input, so it acts as a source and downloads.
 ```yaml
 tasks:
   - name: pull_from_client
     type: sftp
-    operation: download
     host: sftp.client.example.com
     username: '{{ secret "/data/sftp/clientX/username" }}'
     password: '{{ secret "/data/sftp/clientX/password" }}'
@@ -117,26 +112,10 @@ tasks:
     path: s3://my-bucket/inbound/{{ context "CATERPILLAR_FILE_NAME_WRITE" }}
 ```
 
-### List, move, and delete
-```yaml
-tasks:
-  - name: list_incoming
-    type: sftp
-    operation: list
-    host: sftp.client.example.com
-    username: '{{ secret "/data/sftp/clientX/username" }}'
-    password: '{{ secret "/data/sftp/clientX/password" }}'
-    host_key: '{{ secret "/data/sftp/clientX/host_key" }}'
-    remote_path: /incoming
-  - name: echo
-    type: echo
-    only_data: true
-```
-
 ## Notes and limitations
 
 **Files are held in memory.** In the `file → sftp` flow, the `file` task reads each file fully into memory before the `sftp` task sends it. Memory use is therefore set by the size of the largest single file, not the total of all files. This is fine for small and medium files. For very large files (several GB), the task may use too much memory. End-to-end streaming is not supported yet.
 
 **`task_concurrency` opens multiple connections.** Setting `task_concurrency` above `1` makes the task open one SSH connection per worker and transfer files in parallel. This is faster, but it increases memory use (more files in memory at once), and some servers limit the number of connections per user.
 
-**Nested directories are not preserved.** Each file is identified by its base name only. If you upload files from nested S3 folders with a recursive glob (for example `s3://bucket/data/**/*.csv`), they all land in the single `remote_path` directory, and files with the same name overwrite each other. To keep a folder structure, use a separate `file → sftp` pair for each folder, with the target path set for each one.
+**Nested directories are not preserved.** Each file is identified by its base name only. If you download files from nested folders with a recursive glob (for example `/data/**/*.csv`), they all land in the single destination directory, and files with the same name overwrite each other. To keep a folder structure, use a separate `file → sftp` pair for each folder, with the target path set for each one.
