--- test/pipelines/sftp/delete.yaml
@@ -1,13 +1,0 @@
-# Delete a file on the SFTP server.
-tasks:
-  - name: delete_remote_file
-    type: sftp
-    operation: delete
-    host: '{{ secret "/sftp/host" }}'
-    port: 22
-    username: '{{ secret "/sftp/username" }}'
-    password: '{{ secret "/sftp/password" }}'
-    # host_key verifies the server's identity (prevents man-in-the-middle).
-    # Get the value once with: ssh-keyscan -t ed25519 -p <port> <host>
-    host_key: '{{ secret "/sftp/host_key" }}'
-    remote_path: /upload/archive/birds.txt
