+++ test/pipelines/sftp/download.yaml
@@ -1,11 +1,10 @@
 # Download file(s) from an SFTP server to local disk.
-# The file task writes to `path` exactly as given. The sftp download stores
-# each file's name in CATERPILLAR_FILE_NAME_WRITE, so we add it to the path
-# to keep the original file names.
+# The sftp task has no upstream (no input), so it acts as a source -> download.
+# The file task writes to `path` exactly as given, so we add the source file
+# name (stored by the download in CATERPILLAR_FILE_NAME_WRITE) to keep names.
 tasks:
   - name: download_from_sftp
     type: sftp
-    operation: download
     host: '{{ secret "/sftp/host" }}'
     port: 22
     username: '{{ secret "/sftp/username" }}'
