--- test/pipelines/sftp/list.yaml
@@ -1,16 +1,0 @@
-# List a remote directory; echo each entry as JSON.
-tasks:
-  - name: list_remote
-    type: sftp
-    operation: list
-    host: '{{ secret "/sftp/host" }}'
-    port: 22
-    username: '{{ secret "/sftp/username" }}'
-    password: '{{ secret "/sftp/password" }}'
-    # host_key verifies the server's identity (prevents man-in-the-middle).
-    # Get the value once with: ssh-keyscan -t ed25519 -p <port> <host>
-    host_key: '{{ secret "/sftp/host_key" }}'
-    remote_path: /upload
-  - name: echo
-    type: echo
-    only_data: true
