--- test/pipelines/sftp/move.yaml
@@ -1,14 +1,0 @@
-# Rename/move a file on the SFTP server.
-tasks:
-  - name: move_remote_file
-    type: sftp
-    operation: move
-    host: '{{ secret "/sftp/host" }}'
-    port: 22
-    username: '{{ secret "/sftp/username" }}'
-    password: '{{ secret "/sftp/password" }}'
-    # host_key verifies the server's identity (prevents man-in-the-middle).
-    # Get the value once with: ssh-keyscan -t ed25519 -p <port> <host>
-    host_key: '{{ secret "/sftp/host_key" }}'
-    remote_path: /upload/birds.txt
-    destination_path: /upload/archive/birds.txt
