+++ test/pipelines/sftp/upload.yaml
@@ -1,11 +1,12 @@
 # Read a local file and upload it to an SFTP server.
+# The sftp task is placed AFTER a source, so it has an input and acts as a
+# sink -> upload (role is inferred from the channels, like the file task).
 tasks:
   - name: read_local_file
     type: file
     path: test/pipelines/birds.txt
   - name: upload_to_sftp
     type: sftp
-    operation: upload
     host: '{{ secret "/sftp/host" }}'
     port: 22
     username: '{{ secret "/sftp/username" }}'
